Improve data consistency when deleting projects.
There is a race condition when deleting a project: since we are
navigating back to the project list, this retrieves the list from the
server. It may return the just deleted project if this request completes
before the delete or the database transaction is not propagated to the
other connections.

This fix issues a new list request once the delete has been confirmed.
It *will* cause duplicate request to be sent to the server.