Fix NULL facility polygons for some cases
We had at least one case of pgr_alphaShape() returning a polygon with a
hole but without marking it as an inner ring. This is considered invalid
by ST_BuildArea() and resulted in a NULL isochrone.

Use a slightly modified version of pgr_pointsAsPolygon instead which
returns polygon directly instead of passing it through ST_BuildArea if
not necessary.