Keep tasks collection in job when restoring them as pending tasks
The tasks list represents the tasks currently in execution from an import job's standpoint. Thus, when restoring the job, the ongoing tasks are copied (and not renamed) to pending tasks; so the next time a task is requested, a pending task will be used, but there will still be a set of tasks "on the fly" for fsm guard purposes.