Upsert facilities and facility types
Methods for bulk insert first check if the facility and facility types exist, and insert/update as needed. Insert facilities now returns a hash instead of a vector, indicating the status for each of the facilities requested to be inserted.

Deletion methods now also accept an except key, to delete all facilities or types except those just upserted.