Finish job when there are no facilities to process
Add a transition to the import job FSM in which when a :next message
arrives and there are no tasks and no pending facilities to process,
finishes the import process successfully. Also, run the job finisher
at the end of the next-task function when there are no more tasks to
execute to handle this exact case.