Process unprocessed facilities on a dataset reimport
If a dataset is cancelled and then reimported, ensure those facilities that were not processed in the first run are indeed processed.

Change the return type of insert-facilities so it returns a list of all the new facilities inserted, along with an inserted-status attribute that yields whether it's new, moved, existing or updated.

Fixes #250