fix: Avoid NPE when computing suggestions for improvements (#731)
When providers are located in an area outside of the friction data (most
commonly in a coast) but still fall inside the region contour (because it has
better resolution), the engine will fail to compute any demand information for
it. This lead to and NPE in the suggestions algorithm when computing the
relative cost of such providers.